% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_distribution.R
\name{age_distribution}
\alias{age_distribution}
\title{Slice profession and year}
\usage{
age_distribution(data, prof_code, time)
}
\arguments{
\item{data}{A data frame}

\item{profession_code}{profession_code}

\item{year}{year}
}
\value{
A sliced data frame
}
\description{
Slice profession and year
}
