% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_workforce_data.R
\name{aggregate_fte}
\alias{aggregate_fte}
\title{Aggregate workforce data across age groups}
\usage{
aggregate_fte(data)
}
\arguments{
\item{df}{A data frame}
}
\value{
A aggregated data frame
}
\description{
Aggregate workforce data across age groups
}
