% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_workforce_data.r
\name{clean_workforce_data}
\alias{clean_workforce_data}
\title{Clean workforce data}
\usage{
clean_workforce_data(data)
}
\arguments{
\item{df}{A data frame}
}
\value{
A cleaned data frame
}
\description{
Clean workforce data
}
