% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_workforce_data.R
\name{load_workforce_data}
\alias{load_workforce_data}
\title{Loads workforce data}
\usage{
load_workforce_data(path)
}
\arguments{
\item{path}{A string representing the file path to the workforce data CSV file.}
}
\value{
A data frame
}
\description{
Loads workforce data
}
